Added support for older games/gameless days
The library will no longer throw an unhandled exception if no games are
present on a given day when using the ScoreboardOnDate class.
Additionally, team stolen bases and team home runs were removed for
backwards compatibility. The way game start times are read in was also
altered for backwards compatibility.